refactor(Input): make Group style tree shakeable
