+++ components/input/Group.tsx
@@ -6,7 +6,7 @@ import { devUseWarning } from '../_util/warning';
 import { ConfigContext } from '../config-provider';
 import type { FormItemStatusContextProps } from '../form/context';
 import { FormItemInputContext } from '../form/context';
-import useStyle from './style';
+import useStyle from './style/group';
 
 export interface GroupProps {
   className?: string;
